+++ src/app/cart/cart.module.ts
@@ -1,5 +1,4 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { CartListComponent } from './components/cart-list/cart-list.component';
 import { SharedModule } from '../shared/shared.module';
 import { CartItemComponent } from './components/cart-item/cart-item.component';
@@ -12,7 +11,6 @@ import { CartItemComponent } from './components/cart-item/cart-item.component';
     CartItemComponent
   ],
   imports: [
-    CommonModule,
     SharedModule
   ],
   exports: [CartListComponent]