+++ src/app/products/products.module.ts
@@ -1,5 +1,4 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { ProductListComponent } from './product-list/product-list.component';
 import { FirstComponent } from './first/first.component';
 import { SharedModule } from '../shared/shared.module';
@@ -12,7 +11,6 @@ import { SharedModule } from '../shared/shared.module';
     FirstComponent
   ],
   imports: [
-    CommonModule,
     SharedModule
   ],
   exports: [