+++ src/app/shared/pipes/order-by.pipe.ts
@@ -1,0 +1,36 @@
+import { Pipe, PipeTransform } from '@angular/core';
+
+@Pipe({
+  name: 'orderBy'
+})
+export class OrderByPipe implements PipeTransform {
+
+  transform(array: any[], key: string, isAsc: boolean = true): any[] {
+    if(!Array.isArray(array) || array.length <= 1){
+      return array;
+    }
+
+    const compare = (a: any, b: any): number =>{
+      const valueA = this.getItemByPath(a, key);
+      const valueB = this.getItemByPath(b, key);;
+
+      if(valueA < valueB){
+        return isAsc ? -1 : 1;
+      }
+      else if(valueA>valueB){
+        return isAsc ? 1 : -1;
+      }
+      else{
+        return 0;
+      }
+    };
+
+    const clonedArray = [...array];
+    return clonedArray.sort(compare);
+  }
+
+   getItemByPath(obj: any, itemPath: string): any | undefined {
+    return itemPath.split('.').reduce((currentItem, key) => currentItem?.[key], obj);
+  }
+
+}