+++ changelog.md
@@ -52,5 +52,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - applyStyles directive
 
 
+## [0.4.0] - 2023-09-24
+
+### Added
+     - order-by pipe
+     - formsModule and commonModule import and export in shared-module 
+     - sorting and ordering for cart-list component
+     - angular built-in pipe usages like 'currency', 'uppercase', 'async'
+
+ ### Removed
+     - imports and exports of formsModule and commonModule